Refactor: Files refactor
