+++ src/main/java/com/integrador/svfapi/controllers/StudentController.java
@@ -1,6 +1,6 @@
 package com.integrador.svfapi.controllers;
 
-import com.integrador.svfapi.service.StudentServiceIMPL;
+import com.integrador.svfapi.service.impl.StudentServiceIMPL;
 import jakarta.validation.constraints.NotBlank;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
